+++ lib/openmaize/plugs/login.ex
@@ -2,7 +2,7 @@ defmodule Openmaize.Login do
   @moduledoc """
   Plug to handle login.
 
-  There are three options:
+  There are four options:
 
   * redirects - if true, which is the default, redirect on login
   * storage - storage method for the token
@@ -10,6 +10,8 @@ defmodule Openmaize.Login do
     * if storage is set to nil, redirects is automatically set to false
   * token_validity - length of validity of token (in minutes)
     * the default is 1440 minutes (one day)
+  * unique_id - the name which is used to identify the user (in the database)
+    * the default is `:name`
 
   ## Examples with Phoenix
 
@@ -27,26 +29,26 @@ defmodule Openmaize.Login do
 
       plug Openmaize.Login, [storage: nil] when action in [:login_user]
 
-  If you want to store the token in sessionStorage and have the token valid
+  If you want to use `email` to identify the user and have the token valid
   for just two hours:
 
-      plug Openmaize.Login, [storage: nil, token_validity: 120] when action in [:login_user]
+      plug Openmaize.Login, [token_validity: 120, unique_id: :email] when action in [:login_user]
 
   """
 
-  import Ecto.Query
-  import Openmaize.Report
-  import Openmaize.Token
-  alias Openmaize.Config
+  import Openmaize.{Report, Token}
+  alias Openmaize.LoginTools
 
   @behaviour Plug
 
   def init(opts) do
-    token_opts = {0, Keyword.get(opts, :token_validity, 1440)}
-    case Keyword.get(opts, :storage, :cookie) do
-      :cookie -> {Keyword.get(opts, :redirects, true), :cookie, token_opts}
-      nil -> {false, nil, token_opts}
-    end
+    {redirects, storage} = case Keyword.get(opts, :storage, :cookie) do
+             :cookie -> {Keyword.get(opts, :redirects, true), :cookie}
+             nil -> {false, nil}
+           end
+    {redirects, storage, {0, Keyword.get(opts, :token_validity, 1440)},
+     Keyword.get(opts, :unique_id, :name),
+     Keyword.get(opts, :database_call, &LoginTools.check_user/3)}
   end
 
   @doc """
@@ -54,37 +56,22 @@ defmodule Openmaize.Login do
 
   If the login is successful, a JSON Web Token will be returned.
   """
-  def call(%Plug.Conn{params: %{"user" => user_params}} = conn, opts) do
-    user_params |> find_user(Config.unique_id) |> handle_auth(conn, opts)
+  def call(%Plug.Conn{params: %{"user" => user_params}} = conn,
+           {redirects, storage, token_opts, uniq, db_call}) do
+    db_call.(uniq, to_string(uniq), user_params)
+    |> handle_auth(conn, {redirects, storage, token_opts, uniq})
   end
 
-  defp find_user(user_params, uniq) do
-    user = Map.get(user_params, uniq)
-    password = Map.get(user_params, "password")
-    uniq |> String.to_atom |> check_user(user, password)
-  end
-  defp check_user(uniq, user, password) do
-    from(u in Config.user_model,
-         where: field(u, ^uniq) == ^user,
-         select: u)
-    |> Config.repo.one
-    |> check_pass(password)
-  end
-
-  defp check_pass(nil, _), do: Config.get_crypto_mod.dummy_checkpw
-  defp check_pass(%{confirmed: false}, _),
-    do: {:error, "You have to confirm your email address before continuing."}
-  defp check_pass(user, password) do
-    Config.get_crypto_mod.checkpw(password, user.password_hash) and user
-  end
-
-  defp handle_auth(false, conn, {redirects, _, _}) do
+  @doc """
+  Either call the function to create the token or handle the error.
+  """
+  def handle_auth(false, conn, {redirects, _, _, _}) do
     handle_error(conn, "Invalid credentials", redirects)
   end
-  defp handle_auth({:error, message}, conn, {redirects, _, _}) do
+  def handle_auth({:error, message}, conn, {redirects, _, _, _}) do
     handle_error(conn, message, redirects)
   end
-  defp handle_auth(user, conn, opts) do
+  def handle_auth(user, conn, opts) do
     add_token(conn, user, opts)
   end
 end