+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## v0.12.0-dev
+## v0.12.0
 
+* Enhancements
+    * Added option to call custom function to access the database with Openmaize.Login.
+* Changes
+    * Moved the unique_id config value to an option for the Openmaize.Login login function.
 * Bug fixes
     * Fixed the bug with unique_id being set to name in the generate_token funcion.